Externalized input stream to be able to reuse, rewrite, and more easily test that
It was planned to rewrite that, not done yet
