+++ src/main/java/org/htmlunit/cyberneko/HTMLScanner.java
@@ -16,7 +16,6 @@
 package org.htmlunit.cyberneko;
 
 import java.io.EOFException;
-import java.io.FilterInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
@@ -25,6 +24,7 @@ import java.io.UnsupportedEncodingException;
 import java.net.URL;
 import java.util.Locale;
 
+import org.htmlunit.cyberneko.io.PlaybackInputStream;
 import org.htmlunit.cyberneko.util.MiniStack;
 import org.htmlunit.cyberneko.xerces.util.EncodingMap;
 import org.htmlunit.cyberneko.xerces.util.NamespaceSupport;
@@ -3348,216 +3348,6 @@ public class HTMLScanner implements XMLDocumentScanner, XMLLocator, HTMLComponen
             if (DEBUG_BUFFER) {
                 fCurrentEntity.debugBufferIfNeeded(")scanCharacters: ");
             }
-        }
-    }
-
-    /**
-     * A playback input stream. This class has the ability to save the bytes read
-     * from the underlying input stream and play the bytes back later. This class is
-     * used by the HTML scanner to switch encodings when a &lt;meta&gt; tag is
-     * detected that specifies a different encoding.
-     * <p>
-     * If the encoding is changed, then the scanner calls the <code>playback</code>
-     * method and re-scans the beginning of the HTML document again. This should not
-     * be too much of a performance problem because the &lt;meta&gt; tag appears at
-     * the beginning of the document.
-     * <p>
-     * If the &lt;body&gt; tag is reached without playing back the bytes, then the
-     * buffer can be cleared by calling the <code>clear</code> method. This stops
-     * the buffering of bytes and allows the memory used by the buffer to be
-     * reclaimed.
-     * <p>
-     * <strong>Note:</strong> If the buffer is never played back or cleared, this
-     * input stream will continue to buffer the entire stream. Therefore, it is very
-     * important to use this stream correctly.
-     *
-     * @author Andy Clark
-     */
-    private static final class PlaybackInputStream extends FilterInputStream {
-
-        /** Set to true to debug playback. */
-        private static final boolean DEBUG_PLAYBACK = false;
-
-        /** Playback mode. */
-        private boolean playback_;
-
-        /** Buffer cleared. */
-        private boolean cleared_;
-
-        /** Encoding detected. */
-        private boolean detected_;
-
-        // buffer info
-
-        /** Byte buffer. */
-        private byte[] byteBuffer_ = new byte[1024];
-
-        /** Offset into byte buffer during playback. */
-        private int byteOffset_;
-
-        /** Length of bytes read into byte buffer. */
-        private int byteLength_;
-
-        /** Pushback offset. */
-        private int pushbackOffset_;
-
-        /** Pushback length. */
-        private int pushbackLength_;
-
-        // Constructor.
-        PlaybackInputStream(final InputStream in) {
-            super(in);
-        }
-
-        // Detect encoding.
-        public void detectEncoding(final String[] encodings) throws IOException {
-            if (detected_) {
-                throw new IOException("Should not detect encoding twice.");
-            }
-            detected_ = true;
-            final int b1 = read();
-            if (b1 == -1) {
-                return;
-            }
-            final int b2 = read();
-            if (b2 == -1) {
-                pushbackLength_ = 1;
-                return;
-            }
-            // UTF-8 BOM: 0xEFBBBF
-            if (b1 == 0xEF && b2 == 0xBB) {
-                final int b3 = read();
-                if (b3 == 0xBF) {
-                    pushbackOffset_ = 3;
-                    encodings[0] = "UTF-8";
-                    encodings[1] = "UTF8";
-                    return;
-                }
-                pushbackLength_ = 3;
-            }
-            // UTF-16 LE BOM: 0xFFFE
-            if (b1 == 0xFF && b2 == 0xFE) {
-                encodings[0] = "UTF-16";
-                encodings[1] = "UnicodeLittleUnmarked";
-                return;
-            }
-            // UTF-16 BE BOM: 0xFEFF
-            else if (b1 == 0xFE && b2 == 0xFF) {
-                encodings[0] = "UTF-16";
-                encodings[1] = "UnicodeBigUnmarked";
-                return;
-            }
-            // unknown
-            pushbackLength_ = 2;
-        }
-
-        /** Playback buffer contents. */
-        public void playback() {
-            playback_ = true;
-        }
-
-        /**
-         * Clears the buffer.
-         * <p>
-         * <strong>Note:</strong> The buffer cannot be cleared during playback.
-         * Therefore, calling this method during playback will not do anything. However,
-         * the buffer will be cleared automatically at the end of playback.
-         */
-        public void clear() {
-            if (!playback_) {
-                cleared_ = true;
-                byteBuffer_ = null;
-            }
-        }
-
-        /** Read a byte. */
-        @Override
-        public int read() throws IOException {
-            if (DEBUG_PLAYBACK) {
-                System.out.println("(read");
-            }
-            if (pushbackOffset_ < pushbackLength_) {
-                return byteBuffer_[pushbackOffset_++];
-            }
-            if (cleared_) {
-                return in.read();
-            }
-            if (playback_) {
-                final int c = byteBuffer_[byteOffset_++];
-                if (byteOffset_ == byteLength_) {
-                    cleared_ = true;
-                    byteBuffer_ = null;
-                }
-                if (DEBUG_PLAYBACK) {
-                    System.out.println(")read -> " + (char) c);
-                }
-                return c;
-            }
-            final int c = in.read();
-            if (c != -1) {
-                if (byteLength_ == byteBuffer_.length) {
-                    final byte[] newarray = new byte[byteLength_ + 1024];
-                    System.arraycopy(byteBuffer_, 0, newarray, 0, byteLength_);
-                    byteBuffer_ = newarray;
-                }
-                byteBuffer_[byteLength_++] = (byte) c;
-            }
-            if (DEBUG_PLAYBACK) {
-                System.out.println(")read -> " + (char) c);
-            }
-            return c;
-        }
-
-        /** Read an array of bytes. */
-        @Override
-        public int read(final byte[] array) throws IOException {
-            return read(array, 0, array.length);
-        }
-
-        /** Read an array of bytes. */
-        @Override
-        public int read(final byte[] array, final int offset, int length) throws IOException {
-            if (DEBUG_PLAYBACK) {
-                System.out.println(")read(" + offset + ',' + length + ')');
-            }
-            if (pushbackOffset_ < pushbackLength_) {
-                int count = pushbackLength_ - pushbackOffset_;
-                if (count > length) {
-                    count = length;
-                }
-                System.arraycopy(byteBuffer_, pushbackOffset_, array, offset, count);
-                pushbackOffset_ += count;
-                return count;
-            }
-            if (cleared_) {
-                return in.read(array, offset, length);
-            }
-            if (playback_) {
-                if (byteOffset_ + length > byteLength_) {
-                    length = byteLength_ - byteOffset_;
-                }
-                System.arraycopy(byteBuffer_, byteOffset_, array, offset, length);
-                byteOffset_ += length;
-                if (byteOffset_ == byteLength_) {
-                    cleared_ = true;
-                    byteBuffer_ = null;
-                }
-                return length;
-            }
-            final int count = in.read(array, offset, length);
-            if (count != -1) {
-                if (byteLength_ + count > byteBuffer_.length) {
-                    final byte[] newarray = new byte[byteLength_ + count + 512];
-                    System.arraycopy(byteBuffer_, 0, newarray, 0, byteLength_);
-                    byteBuffer_ = newarray;
-                }
-                System.arraycopy(array, offset, byteBuffer_, byteLength_, count);
-                byteLength_ += count;
-            }
-            if (DEBUG_PLAYBACK) {
-                System.out.println(")read(" + offset + ',' + length + ") -> " + count);
-            }
-            return count;
         }
     }
